+++ src/lib/shmem/src/allocator.rs
@@ -1,27 +1,33 @@
-use std::{marker::PhantomData, ops::Deref};
+#![allow(dead_code)]
 
-use once_cell::sync::OnceCell;
+use numtoa::NumToA;
+
+use lazy_static::lazy_static;
 use shadow_pod::Pod;
 use vasi::VirtualAddressSpaceIndependent;
+use vasi_sync::scmutex::SelfContainedMutex;
 
-use crate::c_bindings;
+use crate::util::*;
 
-/// A typed pointer to shared memory.
-///
-/// The pointer to the underlying data (e.g. as accessed via `deref`) is guaranteed
-/// not to change even if the `ShMemBlock` itself is moved. (Host uses this to safely
-/// cache a lock obtained from a ShMemBlock).
+#[derive(Debug)]
 pub struct ShMemBlock<'allocator, T>
-// T must be Sync, since it will be simultaneously available to multiple threads
-// (and processes).
-// T mut be VirtualAddressSpaceIndependent, since it may be simultaneously
-// mapped into different virtual address spaces.
 where
     T: Sync + VirtualAddressSpaceIndependent,
 {
-    internal: c_bindings::ShMemBlock,
-    allocator: &'allocator Allocator,
-    _phantom: PhantomData<T>,
+    block: *mut crate::shmalloc_impl::Block,
+    phantom: core::marker::PhantomData<&'allocator T>,
+}
+
+impl<'allocator, T> ShMemBlock<'allocator, T>
+where
+    T: Sync + VirtualAddressSpaceIndependent,
+{
+    pub fn serialize(&self) -> ShMemBlockSerialized {
+        let serialized = SHMALLOC.lock().internal.serialize(self.block);
+        ShMemBlockSerialized {
+            internal: serialized,
+        }
+    }
 }
 
 // SAFETY: T is already required to be Sync, and ShMemBlock only exposes
@@ -35,301 +41,381 @@ unsafe impl<'allocator, T> Send for ShMemBlock<'allocator, T> where
 {
 }
 
-impl<'allocator, T> ShMemBlock<'allocator, T>
-where
-    T: Sync + VirtualAddressSpaceIndependent,
-{
-    /// Panics if the ShMemBlock is NULL, the incorrect size, or unaligned.
-    ///
-    /// SAFETY: The memory pointed to by `internal` should already be initialized
-    /// to an instance of `T`, and must not be deallocated while the returned
-    /// object is still alive.
-    unsafe fn new(internal: c_bindings::ShMemBlock, allocator: &'allocator Allocator) -> Self {
-        assert_eq!(internal.nbytes, std::mem::size_of::<T>());
-        assert!(!internal.p.is_null());
-        assert_eq!(internal.p.align_offset(std::mem::align_of::<T>()), 0);
-        Self {
-            internal,
-            allocator,
-            _phantom: PhantomData,
-        }
-    }
-
-    pub fn serialize(&self) -> ShMemBlockSerialized {
-        let serialized = unsafe {
-            c_bindings::shmemallocator_blockSerialize(self.allocator.internal, &self.internal)
-        };
-        ShMemBlockSerialized {
-            internal: serialized,
-        }
-    }
-
-    // We require that T is FFI safe (and hence has a stable layout).
-    // This function will fail to compile if it isn't.
-    #[deny(improper_ctypes_definitions)]
-    extern "C" fn _validate_stable_layout(_: T) {}
-}
-
-impl<'allocator, T> Drop for ShMemBlock<'allocator, T>
-where
-    T: Sync + VirtualAddressSpaceIndependent,
-{
-    fn drop(&mut self) {
-        unsafe {
-            std::ptr::drop_in_place(self.internal.p as *mut T);
-            c_bindings::shmemallocator_free(self.allocator.internal, &mut self.internal as *mut _)
-        }
-    }
-}
-
-impl<'allocator, T> Deref for ShMemBlock<'allocator, T>
+impl<'allocator, T> core::ops::Deref for ShMemBlock<'allocator, T>
 where
     T: Sync + VirtualAddressSpaceIndependent,
 {
     type Target = T;
 
     fn deref(&self) -> &Self::Target {
-        // SAFETY: Validity ensured by [`ShMemBlock::new`].
-        unsafe { (self.internal.p as *const T).as_ref().unwrap() }
+        let block = unsafe { &*self.block };
+        &block.get_ref::<T>()[0]
     }
 }
 
-/// An *alias* for a [`ShMemBlock`], which may live in a different process and
-/// address space.
-pub struct ShMemBlockAlias<'serializer, T>
-// T must be Sync, since it will be simultaneously available to multiple threads
-// (and processes).
-// T mut be VirtualAddressSpaceIndependent, since it may be simultaneously
-// mapped into different virtual address spaces.
+impl<'allocator, T> core::ops::Drop for ShMemBlock<'allocator, T>
 where
     T: Sync + VirtualAddressSpaceIndependent,
 {
-    internal: c_bindings::ShMemBlock,
-    // TODO: Extend Serializer to allow cleaning up individual deserialized
-    // blocks, and do so from this objects `drop` method.
-    #[allow(unused)]
-    serializer: &'serializer Serializer,
-    _phantom: PhantomData<T>,
+    fn drop(&mut self) {
+        if !self.block.is_null() {
+            // Guard here to prevent deadlock on free.
+            SHMALLOC.lock().internal.dealloc(self.block);
+            self.block = core::ptr::null_mut();
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct ShMemBlockAlias<'deserializer, T>
+where
+    T: Sync + VirtualAddressSpaceIndependent,
+{
+    block: *mut crate::shmalloc_impl::Block,
+    phantom: core::marker::PhantomData<&'deserializer T>,
 }
 
 // SAFETY: T is already required to be Sync, and ShMemBlock only exposes
 // immutable references to the underlying data.
-unsafe impl<'serializer, T> Sync for ShMemBlockAlias<'serializer, T> where
+unsafe impl<'deserializer, T> Sync for ShMemBlockAlias<'deserializer, T> where
     T: Sync + VirtualAddressSpaceIndependent
 {
 }
-unsafe impl<'serializer, T> Send for ShMemBlockAlias<'serializer, T> where
+unsafe impl<'deserializer, T> Send for ShMemBlockAlias<'deserializer, T> where
     T: Send + Sync + VirtualAddressSpaceIndependent
 {
 }
 
-impl<'serializer, T> ShMemBlockAlias<'serializer, T>
-where
-    T: Sync + VirtualAddressSpaceIndependent,
-{
-    /// Panics if the ShMemBlock is NULL, the incorrect size, or unaligned.
-    ///
-    /// SAFETY: The memory pointed to by `internal` should already be initialized
-    /// to an instance of `T`, and must not be deallocated while the returned
-    /// object is still alive.
-    unsafe fn new(internal: c_bindings::ShMemBlock, serializer: &'serializer Serializer) -> Self {
-        assert_eq!(internal.nbytes, std::mem::size_of::<T>());
-        assert!(!internal.p.is_null());
-        assert_eq!(internal.p.align_offset(std::mem::align_of::<T>()), 0);
-        Self {
-            internal,
-            serializer,
-            _phantom: PhantomData,
-        }
-    }
-
-    // We require that T is FFI safe (and hence has a stable layout).
-    // This function will fail to compile if it isn't.
-    #[deny(improper_ctypes_definitions)]
-    extern "C" fn _validate_stable_layout(_: T) {}
-}
-
-impl<'origin, T> Drop for ShMemBlockAlias<'origin, T>
-where
-    T: Sync + VirtualAddressSpaceIndependent,
-{
-    fn drop(&mut self) {
-        // No cleanup of individual blocks implemented.
-        // TODO: Probably ought to munmap.
-    }
-}
-
-impl<'origin, T> Deref for ShMemBlockAlias<'origin, T>
+impl<'deserializer, T> core::ops::Deref for ShMemBlockAlias<'deserializer, T>
 where
     T: Sync + VirtualAddressSpaceIndependent,
 {
     type Target = T;
 
     fn deref(&self) -> &Self::Target {
-        // SAFETY: Validity ensured by [`ShMemBlockAlias::new`].
-        unsafe { (self.internal.p as *const T).as_ref().unwrap() }
+        let block = unsafe { &*self.block };
+        &block.get_ref::<T>()[0]
     }
 }
 
-/// A serialized descriptor for a `ShMemBlock`, suitable to be transferred
-/// across processes, which can be used to create a `ShMemBlockAlias` referencing
-/// the original `ShMemBlock`.
 #[derive(Copy, Clone, Debug, VirtualAddressSpaceIndependent)]
 #[repr(transparent)]
 pub struct ShMemBlockSerialized {
-    internal: c_bindings::ShMemBlockSerialized,
+    internal: crate::shmalloc_impl::BlockSerialized,
 }
+
 unsafe impl Pod for ShMemBlockSerialized {}
 
-// SAFETY: This is a serialized blob, designed to be VASI.
-unsafe impl VirtualAddressSpaceIndependent for c_bindings::ShMemBlockSerialized {}
-
 impl ShMemBlockSerialized {
-    // Keep in sync with macro of same name in shmem_allocator.h.
-    const SHD_SHMEM_BLOCK_SERIALIZED_MAX_STRLEN: usize = 21 + 21 + 21 + 256 + 1;
+    pub fn to_string_buf(self) -> StringBuf {
+        let mut retval = StringBuf([0; STRING_BUF_NBYTES]);
 
-    pub fn encode_to_string(&self) -> String {
-        let mut buf = Vec::new();
-        buf.resize(Self::SHD_SHMEM_BLOCK_SERIALIZED_MAX_STRLEN, 0i8);
-        unsafe { c_bindings::shmemblockserialized_toString(&self.internal, buf.as_mut_ptr()) };
-        let buf = buf
-            .iter()
-            .take_while(|c| **c != 0)
-            .map(|c| *c as u8)
-            .collect();
-        String::from_utf8(buf).unwrap()
+        let mut offset_buf = [0u8; 32];
+        self.internal.offset.numtoa(10, &mut offset_buf);
+
+        let i1 = offset_buf.iter().filter(|x| **x != 0);
+        let i2 = ";".as_bytes().iter();
+        let i3 = self.internal.chunk_name.iter();
+
+        let i_all = i1.chain(i2).chain(i3);
+
+        retval.0.iter_mut().zip(i_all).for_each(|(x, y)| *x = *y);
+
+        retval
     }
 
-    pub fn decode_from_string(s: &str) -> Option<Self> {
-        let mut err: bool = false;
-        let mut buf: Vec<i8> = s.as_bytes().iter().map(|b| *b as i8).collect();
-        // Null terminate.
-        buf.push(0);
-        let res = Self {
-            internal: unsafe {
-                c_bindings::shmemblockserialized_fromString(buf.as_ptr(), &mut err)
-            },
+    pub fn from_string_buf(string_buf: StringBuf) -> Self {
+        const DELIM: u8 = 59; // Decimal value of ;
+
+        let lhs_itr = string_buf.0.iter();
+        let mut rhs_itr = string_buf.0.iter();
+        rhs_itr.find(|x| **x == DELIM);
+
+        let offset_itr = lhs_itr.take_while(|x| **x != DELIM);
+
+        let mut offset_buf = [0u8; 32];
+        offset_buf
+            .iter_mut()
+            .zip(offset_itr)
+            .for_each(|(x, y)| *x = *y);
+
+        let mut path_buf: crate::shmalloc_impl::PathBuf =
+            [0; crate::shmalloc_impl::PATH_BUF_NBYTES];
+
+        path_buf.iter_mut().zip(rhs_itr).for_each(|(x, y)| *x = *y);
+
+        // Unwraps safe here.
+        let end = offset_buf.iter().position(|&x| x == 0).unwrap();
+
+        let offset = unsafe {
+            core::str::from_utf8_unchecked(&offset_buf[0..end])
+                .parse::<isize>()
+                .unwrap()
         };
-        if err {
-            None
-        } else {
-            Some(res)
+
+        ShMemBlockSerialized {
+            internal: crate::shmalloc_impl::BlockSerialized {
+                chunk_name: path_buf,
+                offset,
+            },
         }
     }
 }
 
-/// An allocator for `ShMemBlock`s.
-pub struct Allocator {
-    internal: *mut c_bindings::ShMemAllocator,
+lazy_static! {
+    pub static ref SHMALLOC: SelfContainedMutex<SharedMemAllocator<'static>> = {
+        let alloc = SharedMemAllocator::new();
+        SelfContainedMutex::new(alloc)
+    };
+    pub static ref SHDESERIALIZER: SelfContainedMutex<SharedMemDeserializer<'static>> = {
+        let deserial = SharedMemDeserializer::new();
+        SelfContainedMutex::new(deserial)
+    };
 }
 
-// SAFETY: The C bindings for ShMemAllocator use internal locking to ensure
-// thread-safe access.
-//
-// TODO: Consider using SyncSendPointer instead of directly implementing these.
-// Would require reorganizing crates to avoid a circular dependency, though.
-unsafe impl Send for Allocator {}
-unsafe impl Sync for Allocator {}
+pub fn shmalloc<T>(val: T) -> ShMemBlock<'static, T>
+where
+    T: Sync + VirtualAddressSpaceIndependent,
+{
+    SHMALLOC.lock().alloc(val)
+}
 
-impl Allocator {
-    pub fn global() -> &'static Self {
-        static GLOBAL: OnceCell<Allocator> = OnceCell::new();
-        GLOBAL.get_or_init(|| Self {
-            internal: unsafe { c_bindings::shmemallocator_getGlobal() },
-        })
+pub fn shfree<T>(block: ShMemBlock<'static, T>)
+where
+    T: Sync + VirtualAddressSpaceIndependent,
+{
+    SHMALLOC.lock().free(block);
+}
+
+pub unsafe fn shdeserialize<T>(serialized: &ShMemBlockSerialized) -> ShMemBlockAlias<'static, T>
+where
+    T: Sync + VirtualAddressSpaceIndependent,
+{
+    unsafe { SHDESERIALIZER.lock().deserialize(serialized) }
+}
+
+// lazy_static wants these.
+
+unsafe impl Send for SharedMemAllocator<'_> {}
+unsafe impl Sync for SharedMemAllocator<'_> {}
+
+unsafe impl Send for SharedMemDeserializer<'_> {}
+unsafe impl Sync for SharedMemDeserializer<'_> {}
+
+pub struct SharedMemAllocator<'alloc> {
+    internal: crate::shmalloc_impl::FreelistAllocator,
+    nallocs: isize,
+    phantom: core::marker::PhantomData<&'alloc ()>,
+}
+
+impl<'alloc> SharedMemAllocator<'alloc> {
+    fn new() -> Self {
+        let mut internal = crate::shmalloc_impl::FreelistAllocator::new();
+        internal.init().unwrap();
+
+        Self {
+            internal,
+            nallocs: 0,
+            phantom: Default::default(),
+        }
     }
 
-    pub fn alloc<T>(&self, val: T) -> ShMemBlock<T>
-    where
-        T: Sync + VirtualAddressSpaceIndependent,
-    {
-        let nbytes = std::mem::size_of_val(&val);
-        let raw_block = unsafe { c_bindings::shmemallocator_alloc(self.internal, nbytes) };
-        assert_eq!(raw_block.nbytes, nbytes);
-        assert!(!raw_block.p.is_null());
-        assert_eq!(raw_block.p.align_offset(std::mem::align_of::<T>()), 0);
-        // Safety: We've validated non-null, size, and alignment.
-        unsafe { (raw_block.p as *mut T).write(val) };
-        // Safety: We've correctly initialized the raw_block, and are
-        // transferring sole ownership to the ShMemBlock.
-        unsafe { ShMemBlock::new(raw_block, self) }
+    fn alloc<T: Sync + VirtualAddressSpaceIndependent>(&mut self, val: T) -> ShMemBlock<'alloc, T> {
+        let t_nbytes: usize = core::mem::size_of::<T>();
+        let t_alignment: usize = core::mem::align_of::<T>();
+
+        let block = self.internal.alloc(t_nbytes, t_alignment);
+        unsafe {
+            (*block).get_mut_ref::<T>()[0] = val;
+        }
+
+        self.nallocs += 1;
+        ShMemBlock::<'alloc, T> {
+            block,
+            phantom: Default::default(),
+        }
+    }
+
+    fn free<T: Sync + VirtualAddressSpaceIndependent>(&mut self, mut block: ShMemBlock<'alloc, T>) {
+        self.nallocs -= 1;
+        block.block = core::ptr::null_mut();
+        self.internal.dealloc(block.block);
+    }
+
+    fn destruct(&mut self) {
+        self.internal.destruct();
+
+        if self.nallocs != 0 {
+            crate::shmalloc_impl::log_error(crate::shmalloc_impl::AllocError::Leak, None);
+        }
     }
 }
 
-/// Transforms `ShMemBlockSerialized` to `ShMemBlockAlias`.
-pub struct Serializer {
-    internal: *mut crate::c_bindings::ShMemSerializer,
+// Experimental... implements the global allocator using the shared memory allocator.
+/*
+unsafe impl Sync for GlobalAllocator {}
+unsafe impl Send for GlobalAllocator {}
+
+struct GlobalAllocator {}
+
+unsafe impl core::alloc::GlobalAlloc for GlobalAllocator {
+    unsafe fn alloc(&self, layout: core::alloc::Layout) -> *mut u8 {
+        let block_p = SHMALLOC
+            .lock()
+            .internal
+            .alloc(layout.size(), layout.align());
+        let (p, _) = unsafe { (*block_p).get_mut_block_data_range() };
+        p
+    }
+
+    unsafe fn dealloc(&self, ptr: *mut u8, _layout: core::alloc::Layout) {
+        let block_p = crate::shmalloc_impl::rewind(ptr);
+        SHMALLOC.lock().internal.dealloc(block_p);
+    }
+}
+*/
+
+pub struct SharedMemDeserializer<'alloc> {
+    internal: crate::shmalloc_impl::FreelistDeserializer,
+    phantom: core::marker::PhantomData<&'alloc ()>,
 }
 
-// SAFETY: The C bindings for ShMemSerializer use internal locking to ensure
-// thread-safe access.
-//
-// TODO: Consider using SyncSendPointer instead of directly implementing these.
-// Would require reorganizing crates to avoid a circular dependency, though.
-unsafe impl Send for Serializer {}
-unsafe impl Sync for Serializer {}
+impl<'alloc> SharedMemDeserializer<'alloc> {
+    fn new() -> Self {
+        let internal = crate::shmalloc_impl::FreelistDeserializer::new();
 
-impl Serializer {
-    pub fn global() -> &'static Self {
-        static GLOBAL: OnceCell<Serializer> = OnceCell::new();
-        GLOBAL.get_or_init(|| Self {
-            internal: unsafe { c_bindings::shmemserializer_getGlobal() },
-        })
+        Self {
+            internal,
+            phantom: Default::default(),
+        }
     }
 
     /// # Safety
     ///
-    /// * `block` must have been created from a `ShMemBlock<T>`
-    /// * The returned `ShMemBlockAlias` must not outlive the original `ShMemBlock`
-    ///   that `block` was serialized from. We can't guarantee this with normal
-    ///   lifetime analysis, since the original block may be in another process.
-    pub unsafe fn deserialize<'a, T>(
-        &'a self,
-        block: &ShMemBlockSerialized,
-    ) -> ShMemBlockAlias<'a, T>
+    /// This function can violate type safety if a template type is provided that does not match
+    /// original block that was serialized.
+    pub unsafe fn deserialize<T>(
+        &mut self,
+        serialized: &ShMemBlockSerialized,
+    ) -> ShMemBlockAlias<'alloc, T>
     where
         T: Sync + VirtualAddressSpaceIndependent,
     {
-        let raw_blk =
-            unsafe { c_bindings::shmemserializer_blockDeserialize(self.internal, &block.internal) };
-        unsafe { ShMemBlockAlias::new(raw_blk, self) }
+        let block = self.internal.deserialize(&serialized.internal);
+
+        ShMemBlockAlias {
+            block,
+            phantom: Default::default(),
+        }
+    }
+}
+
+struct SharedMemAllocatorDropGuard;
+
+impl Drop for SharedMemAllocatorDropGuard {
+    fn drop(&mut self) {
+        SHMALLOC.lock().destruct();
     }
 }
 
 #[cfg(test)]
 mod tests {
+    use super::*;
+    use rand::Rng;
     use std::sync::atomic::{AtomicI32, Ordering};
 
-    use super::*;
+    extern "C" fn shmalloc_teardown() {
+        SHMALLOC.lock().destruct();
+    }
+
+    // Just needed because we can't put the drop guard in a global place. We don't want to drop
+    // after every function, and there's no global test main routine we can take advantage of. No
+    // big deal.
+    fn register_teardown() {
+        use std::sync::Mutex;
+        static MTX: Mutex<i32> = Mutex::new(0);
+        let _guard = MTX.lock();
+
+        static mut INIT: bool = false;
+
+        unsafe {
+            if !INIT {
+                libc::atexit(shmalloc_teardown);
+                INIT = true;
+            }
+        }
+    }
 
     #[test]
-    // Uses FFI
+    fn allocator_random_allocations() {
+        register_teardown();
+
+        const NROUNDS: usize = 100;
+        let mut marked_blocks: Vec<(u32, ShMemBlock<u32>)> = Default::default();
+        let mut rng = rand::thread_rng();
+
+        let mut execute_round = || {
+            // Some allocations
+            for i in 0..255 {
+                let b = shmalloc(i);
+                marked_blocks.push((i, b));
+            }
+
+            // Generate some number of items to pop
+            let n1: u8 = rng.gen();
+
+            for _ in 0..n1 {
+                let last_marked_block = marked_blocks.pop().unwrap();
+                assert_eq!(last_marked_block.0, *last_marked_block.1);
+                shfree(last_marked_block.1);
+            }
+
+            // Then check all blocks
+            for idx in 0..marked_blocks.len() {
+                assert_eq!(marked_blocks[idx].0, *marked_blocks[idx].1);
+            }
+        };
+
+        for _ in 0..NROUNDS {
+            execute_round();
+        }
+
+        while marked_blocks.len() > 0 {
+            let b = marked_blocks.pop().unwrap();
+            shfree(b.1);
+        }
+    }
+
     #[cfg_attr(miri, ignore)]
     fn round_trip_through_serializer() {
+        register_teardown();
         type T = i32;
         let x: T = 42;
 
-        let original_block: ShMemBlock<T> = Allocator::global().alloc(x);
+        let original_block: ShMemBlock<T> = shmalloc(x);
         {
             let serialized_block = original_block.serialize();
-            let serialized_str = serialized_block.encode_to_string();
-            let serialized_block =
-                ShMemBlockSerialized::decode_from_string(&serialized_str).unwrap();
-            let block = unsafe { Serializer::global().deserialize::<T>(&serialized_block) };
+            let serialized_str = serialized_block.to_string_buf();
+            let serialized_block = ShMemBlockSerialized::from_string_buf(serialized_str);
+            let block = unsafe { shdeserialize::<i32>(&serialized_block) };
             assert_eq!(*block, 42);
         }
+
+        shfree(original_block);
     }
 
     #[test]
     // Uses FFI
     #[cfg_attr(miri, ignore)]
     fn mutations() {
+        register_teardown();
+
         type T = AtomicI32;
-        let original_block = Allocator::global().alloc(AtomicI32::new(0));
+        let original_block = shmalloc(AtomicI32::new(0));
 
         let serialized_block = original_block.serialize();
-        let deserialized_block =
-            unsafe { Serializer::global().deserialize::<T>(&serialized_block) };
+
+        let deserialized_block = unsafe { shdeserialize::<T>(&serialized_block) };
 
         assert_eq!(original_block.load(Ordering::SeqCst), 0);
         assert_eq!(deserialized_block.load(Ordering::SeqCst), 0);
@@ -343,6 +429,8 @@ mod tests {
         deserialized_block.store(20, Ordering::SeqCst);
         assert_eq!(original_block.load(Ordering::SeqCst), 20);
         assert_eq!(deserialized_block.load(Ordering::SeqCst), 20);
+
+        shfree(original_block);
     }
 
     // Validate our guarantee that the data pointer doesn't move, even if the block does.
@@ -351,21 +439,25 @@ mod tests {
     // Uses FFI
     #[cfg_attr(miri, ignore)]
     fn shmemblock_stable_pointer() {
+        register_teardown();
+
         type T = u32;
-        let block: ShMemBlock<T> = Allocator::global().alloc(0);
+        let original_block: ShMemBlock<T> = shmalloc(0);
 
-        let block_addr = &block as *const ShMemBlock<T>;
-        let data_addr = block.deref() as *const T;
+        let block_addr = &original_block as *const ShMemBlock<T>;
+        let data_addr = *original_block as *const T;
 
-        let block = Some(block);
+        let block = Some(original_block);
 
         // Validate that the block itself actually moved.
         let new_block_addr = block.as_ref().unwrap() as *const ShMemBlock<T>;
         assert_ne!(block_addr, new_block_addr);
 
         // Validate that the data referenced by the block *hasn't* moved.
-        let new_data_addr = block.as_ref().unwrap().deref() as *const T;
+        let new_data_addr = **(block.as_ref().unwrap()) as *const T;
         assert_eq!(data_addr, new_data_addr);
+
+        shfree(block.unwrap());
     }
 
     // Validate our guarantee that the data pointer doesn't move, even if the block does.
@@ -373,21 +465,26 @@ mod tests {
     // Uses FFI
     #[cfg_attr(miri, ignore)]
     fn shmemblockremote_stable_pointer() {
-        type T = u32;
-        let allocd_block: ShMemBlock<T> = Allocator::global().alloc(0);
-        let block = unsafe { Serializer::global().deserialize::<T>(&allocd_block.serialize()) };
+        register_teardown();
 
-        let block_addr = &block as *const _;
-        let data_addr = block.deref() as *const T;
+        type T = u32;
+        let alloced_block: ShMemBlock<T> = shmalloc(0);
+
+        let block = unsafe { shdeserialize::<T>(&alloced_block.serialize()) };
+
+        let block_addr = &block as *const ShMemBlockAlias<T>;
+        let data_addr = *block as *const T;
 
         let block = Some(block);
 
         // Validate that the block itself actually moved.
-        let new_block_addr = block.as_ref().unwrap() as *const _;
+        let new_block_addr = block.as_ref().unwrap() as *const ShMemBlockAlias<T>;
         assert_ne!(block_addr, new_block_addr);
 
         // Validate that the data referenced by the block *hasn't* moved.
-        let new_data_addr = block.as_ref().unwrap().deref() as *const T;
+        let new_data_addr = **(block.as_ref().unwrap()) as *const T;
         assert_eq!(data_addr, new_data_addr);
+
+        shfree(alloced_block);
     }
 }
