+++ src/main/host/host.rs
@@ -239,7 +239,7 @@ impl Host {
             manager_shmem,
         );
         let shim_shmem =
-            UnsafeCell::new(shadow_shmem::allocator::Allocator::global().alloc(host_shmem));
+            UnsafeCell::new(shadow_shmem::allocator::shmalloc(host_shmem));
 
         // Process IDs start at 1000
         let thread_id_counter = Cell::new(1000);
